requirements.txt
# Flask and Extensions
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0

# Database
pymongo==4.6.2
mongomock==4.1.2
redis==5.0.2

# Task Queue
celery==5.3.6
kombu==5.3.5

# Utilities
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0

# Logging
structlog==24.1.0

# Type Hints
typing-extensions==4.10.0

# Development Tools
black==24.2.0
flake8==7.0.0
isort==5.13.2 